[47] - Created InputField as only config dependency and restructured database column list for inheritance, avoiding duplication
